pyproject.toml
[project]
name = "naheal-cancer-wiki"
version = "0.1.0"
description = "자연 치유 LLM Wiki 빌드 스크립트"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0",
    "firebase-admin>=6.0.0",
]

[project.scripts]
compile = "scripts.compile:main"

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
